refactor(ui): implement PR suggestion to make directive
