+++ src/app/ui/common/expand-image-button.html
@@ -1,0 +1,5 @@
+<md-button class="md-icon-button rv-icon-20 rv-image-expand" ng-click="self.onExpandClick(self.block.content)"
+    aria-label="{{ 'toc.tooltip.fullsizeImage' | translate }}">
+    <md-icon md-svg-src="action:zoom_in"></md-icon>
+    <md-tooltip md-direction="top">{{ 'toc.tooltip.fullsizeImage' | translate }}</md-tooltip>
+</md-button>
