+++ src/app/ui/common/expand-image-dialog.html
@@ -1,0 +1,6 @@
+<md-dialog aria-label="{{\'toc.tooltip.fullsizeImage\' | translate}}" class="rv-expanded-dialog">
+    <rv-content-pane close-panel="self.close()" title-style="title" title-value="{{\'toc.tooltip.fullsizeImage\' | translate}}">
+        <img class="rv-expanded-image" src="{{ self.imageUrl }}"></img>
+        <div>{{self.test}}</div>
+    </rv-content-pane>
+</md-dialog>
